requirements.txt
langchain==0.2.4
langchain_chroma==0.1.1
langchain_community==0.2.4
langchain_openai==0.1.8
langchain_text_splitters==0.2.1
openai==1.34.0
python-dotenv==1.0.1

